Added multiple changes pending
There are a few changes, mostly involving polybar that haven't been
staged to the repository yet because the dotfiles haven't been symlinked
in my system yet, so I have to manually modify the files in the
repository in order to account for the changes I'm actually making.

This should be fixed after this commit.