Added support for multiple monitors
The configuration is somewhat hardcoded to my own setup, since it
assumes that your main monitor is labeled 'eDP1' and that any other
secondary monitors are labeled 'HDMI2' or 'DP1'.

This works properly with autorandr, in any setup from one to three
screens. Right now the polybar in all the bars are the same, though it's
easy to modify to make different polybar bars show up in different
monitors.